refactor(unpack): make strings in outputs.yaml less verbose
The default string style uses yaml string literals (|) whenever a string
contains a newline. Since outputs in Jupyter notebooks are (usually)
split line by line, we end up with not-so-compact yaml. This commit
makes it so that whenever we have a yaml string that only has one
newline and that newline is at the end of the string, we use double
quotes instead of string literals.