tests: prevent markdown utility tests from generating newlines in text
This change makes the generator generate alphanumeric text instead of
unicode to avoid generating newlines (we shouldn't generate newlines
because we're generating, well, lines). This is just a band-aid for now
as we completely miss out on unicode text, so I need to revisit this in
the future (probably a custom generator to generate unicode, but without
newlines).