fix(unpack, pack): fix crash on multiline inline html comments in markdown
Seems like Commonmark has multiple source-position pairs in the source
range for multiline inline html comments, which broke my assumption that
the source ranges will only have a single pair. This commit fixes that
by "collapsing" source ranges with multiple source position pairs so
that we get a single pair of start and end indices that covers the whole
effective range.

Unfortunately, Commonmark seems to remove some of the whitespace in the
HTML comment, so we have to refer back to the original full markdown
text to extract the full HTML comment.