refactor(readme): don't mention where cabal installs executables to
It's not clear from Cabal's documentation where it actually installs
executables to with the default configuration. I believe it's
`~/.local/bin` on Unix-like systems, but there's nothing about the
default for Windows. Best to just leave this out from the readme.
Running `cabal install` will output the install path anyway.